Streamline documentation
This is a reorganizing of the Placeholder documentation with the goal
of allowing for new users to quickly understand the need for, and how to
install, Placeholder.

To that end, the initial description and installation instructions have
been made more concise, with the more detailed info moved further down
in the readme.

Additionally, capitalization has been fixed.