Use server start script instead of NPM
As mentioned in https://github.com/pelias/api/pull/1096 and
https://github.com/npm/npm/issues/4603, `npm start` is not the preferred
way to start services in a Docker container. There are now several
reasons:

1. `npm` does not properly handle signals making it harder to stop/kill
containers
2. `npm` attempts to write to disk, making it impossible to use with a
read only root filesystem (which is a good security-concious best
practice for use with containers).

This change creates a server script which starts Placeholder
appropriately, and uses that in the Dockerfile. To avoid duplicating
knowledge of how to start Placeholder, `npm start` uses it as well.