Use NODE_ENV ARG to control NPM install and tests
By using the Dockerfile [ARG](https://docs.docker.com/engine/reference/builder/#arg)
directive, we can set an ENV variable and control it via the
`--build-arg` parameter to `docker build`.

Using this we can configure the Dockerfile to skip installing NPM dev
dependencies and running unit tests.

In my testing this reduces the size of the Placeholder docker images
from 290MB to 200MB, which is pretty huge! It reduces the time to build
the image, reduces our surface area for security issues, and seems like
an all around win.