feat(http): Improve signal handling and logging
This PR improves the signal handling of Placeholder so that it exits
gracefully in all cases, even when running under Docker.

There appear to have been several interconnected issues:

- SIGINT was not handled, which is the signal sent by `ctrl-C` on the
command line and by docker

- the signal handler called `app.close()`, which is not defined. Instead
`.close()` must be called on the http server returned from
`app.listen()`

Because of the `cluster` module, there's a few different paths through
the code: either with multiple CPUs or with a single CPU. In the case of
multiple CPUs the signal handler has to work for both a worker and
master process.

Along the way, logging has been a bit streamlined and now uses
`pelias-logger`.