feat(dirs): Update directories to work with bundles/sqlite
BREAKING_CHANGE: The path used to specify where Who's on First data is
located should now _not_ end with `/data`. This will only affect people
using Placeholder outside of the [Pelias
Docker](https://github.com/pelias/docker) environment.

This PR ensures that the bundle _and_ sqlite code in this repo can share
a common base path set by `WOF_DIR`.

In the past, each required a different directory to work correctly
because they depended on the whosonfirst layout directories of `data`
and `sqlite` to be encoded in the `WOF_DIR` environment variable.

Now, the knowledge of which directory from the standard WOF layout to
use is encoded in the extraction code. The bundle code looks in the
`data` subdirectory, whereas the SQLite code looks in the `sqlite`
directory.

The environment variable override in the `Dockerfile` have been updated
accordingly.