Add query combination limit
Recently, we've noticed some queries sent to Placeholder that take a
very long time to complete (upwards of 30 seconds depending on the CPU
Placeholder is running on).

Because Placeholder is a Node.js process that makes synchronous SQL
queries to SQLite, slow queries are a big problem: they slow down all
other queries being sent to that same Node.js process. This can be
somewhat worked around by the `cluster` module built into Node.js, but
overall slow queries are simply very bad and hard to work around

Causes
======

The ultimate cause of these slow queries is that some of the Placeholder
SQL queries have logic like: "find all records with tokens in this list,
parented by all records with tokens in this other list".

If either or both of those lists of tokens are very common, it's
possible for the number of combinations to explode into the tens of
millions.

Even though the amount of data in a full planet Placeholder build is
fairly small, this can of course end up quite slow.

The Fix
=======

Instead of attempting to write queries that will always work, which is
probably difficult and would require huge changes to the Placeholder
logic, this PR attempts to detect queries that would be too slow, and
prevent them from being run.

In practice, this works out very well, as Pelias in general can handle
empty results coming back from Placeholder if needed. In general
Placeholder will run many similar SQL queries, so there might be results
from another SQL query, even if one is skipped because it would be too
slow.

It's _strictly_ better than the behavior before this PR, which is that
sometimes a very slow query would effectively cause many fast queries to
become slow.