feat(extract): Detect location of whosonfirst data from pelias.json
This change allows the Placeholder extract script to work in most cases
_without_ specifying the `WOF_DIR` environment variable.

Previously, unless you were using the particular arrangement of files
and directories from pelias/docker, the default location the extract script
looks for data (`/data/whosonfirst-data/sqlite`) was probably not
correct.

I noticed this inconvenience when running Pelias locally _without_
docker for the first time in quite a long time.

My guess/recollection is an older version of the extract script
(pre-sqlite) was pure bash, and so checking `pelias.json` was less
convenient than in the current Node.js script.

The `WOF_DIR` environment variable is left as an override, but my hope
is with this change almost no one would have to use it.